adk-demo/invitations/pyproject.toml
[project]
name = "invitations"
version = "0.1.0"
description = ""
authors = [
    {name = "gergely.csonde",email = "gergely.csonde@aliz.ai"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic (>=2.11.7,<3.0.0)",
    "google-adk (>=1.4.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "invitations", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
